refactor(core): split request and test modules
